refactor: modify code style to code reviews
